Handle release PR creation
Webhooks notify the application that a pull request was created.
If a pull request was opened on a branch that is being tracked
by any of the releases, trigger a re-sync for that release.

This allows the state of a release to automatically be updated
to Code Freeze once a release operator uses the Create Release
Action on GitHub on code freeze day.