Changed API to accept releaseId instead of releaseName
While release names are unique, they raise issues when attempting to
modify releases. If we want to modify the name of a release (it's
primary key) we need to pass both the previous release name and the new
release name. A cleaner, more correct approach is to just give clients
the Ids of the releases, and accept those instead. This also removes the
need to lookup the release with the given release name.